+++ Plugins/BEAMPROJ_HathoraDemo/Source/BEAMPROJ_HathoraDemo/BEAMPROJ_HathoraDemo.Build.cs
@@ -13,14 +13,6 @@ public class BEAMPROJ_HathoraDemo : ModuleRules
 			{
 				"Core",
 				
-				// You only need this to access the BeamOSS namespace that has some utility functions for integrating with OnlineSubsystem...
-				// If you don't want it, you can simply remove it and do what the utility functions do in your own project.
-				"BeamableCore",
-				
-				// We have this here because we use the UBeamLevelSubsystem --- if you wish to depend only on the OnlineSubsystem stuff in your game code, you can just remove this dependency
-				// and make your own UBeamLevelSubsystem-equivalent.
-				"BeamableCoreRuntime",
-				
 				"OnlineSubsystem",
 				"OnlineSubsystemUtils",
 				"OnlineSubsystemBeamable",
@@ -37,5 +29,11 @@ public class BEAMPROJ_HathoraDemo : ModuleRules
 				"Slate",
 				"SlateCore",
 			});
+		
+		// You only need this to access the BeamOSS namespace that has some utility functions for integrating with OnlineSubsystem...
+		// If you don't want it, you can simply remove it and do what the utility functions do in your own project.
+		// We have this here because we use the UBeamLevelSubsystem --- if you wish to depend only on the OnlineSubsystem stuff in your game code, you can just remove this dependency
+		// and make your own UBeamLevelSubsystem-equivalent.
+		Beam.AddRuntimeModuleDependencies(this);
 	}
 }