+++ Plugins/BeamableCore/Source/BeamableCoreRuntimeEditor/Public/Subsystems/CLI/Autogen/BeamCliProjectBuildCommand.h
@@ -48,7 +48,7 @@ Usage:
 
 Options:
   -w, --watch                      When true, the command will run forever and watch the state of the program
-  --ids <ids>                      The list of services to build, defaults to all local services
+  --ids <ids>                      The list of services to include, defaults to all local services
   --dryrun                         Should any networking happen?
   --cid <cid>                      Cid to use; will default to whatever is in the file system
   --pid <pid>                      Pid to use; will default to whatever is in the file system