+++ Plugins/BeamableCore/Source/BeamableCoreRuntimeEditor/Public/Subsystems/CLI/Autogen/BeamCliProjectDisableCommand.cpp
@@ -1,0 +1,38 @@
+#include "BeamCliProjectDisableCommand.h"
+
+#include "BeamLogging.h"
+#include "Serialization/JsonSerializerMacros.h"
+
+FString UBeamCliProjectDisableCommand::GetCommand()
+{
+	return FString(TEXT("project disable"));
+}
+		
+void UBeamCliProjectDisableCommand::HandleStreamReceived(FBeamOperationHandle Op, FString ReceivedStreamType, int64 Timestamp, TSharedRef<FJsonObject> DataJson, bool isServer)
+{
+	if(ReceivedStreamType.Equals(StreamType) && OnStreamOutput)
+	{
+		UBeamCliProjectDisableStreamData* Data = NewObject<UBeamCliProjectDisableStreamData>(this);
+		Data->OuterOwner = this;
+		Data->BeamDeserializeProperties(DataJson);
+
+		Stream.Add(Data);
+		Timestamps.Add(Timestamp);
+		
+		AsyncTask(ENamedThreads::GameThread, [this, Op]
+		{
+			OnStreamOutput(Stream, Timestamps, Op);
+		});				
+	}
+}
+
+void UBeamCliProjectDisableCommand::HandleStreamCompleted(FBeamOperationHandle Op, int ResultCode, bool isServer)
+{
+	if (OnCompleted)
+	{
+		AsyncTask(ENamedThreads::GameThread, [this, ResultCode, Op]
+		{
+			OnCompleted(ResultCode, Op);
+		});
+	}
+}