+++ Plugins/BeamableCore/Source/BeamableCoreRuntimeEditor/Public/Subsystems/CLI/Autogen/BeamCliProjectEnableCommand.cpp
@@ -1,0 +1,38 @@
+#include "BeamCliProjectEnableCommand.h"
+
+#include "BeamLogging.h"
+#include "Serialization/JsonSerializerMacros.h"
+
+FString UBeamCliProjectEnableCommand::GetCommand()
+{
+	return FString(TEXT("project enable"));
+}
+		
+void UBeamCliProjectEnableCommand::HandleStreamReceived(FBeamOperationHandle Op, FString ReceivedStreamType, int64 Timestamp, TSharedRef<FJsonObject> DataJson, bool isServer)
+{
+	if(ReceivedStreamType.Equals(StreamType) && OnStreamOutput)
+	{
+		UBeamCliProjectEnableStreamData* Data = NewObject<UBeamCliProjectEnableStreamData>(this);
+		Data->OuterOwner = this;
+		Data->BeamDeserializeProperties(DataJson);
+
+		Stream.Add(Data);
+		Timestamps.Add(Timestamp);
+		
+		AsyncTask(ENamedThreads::GameThread, [this, Op]
+		{
+			OnStreamOutput(Stream, Timestamps, Op);
+		});				
+	}
+}
+
+void UBeamCliProjectEnableCommand::HandleStreamCompleted(FBeamOperationHandle Op, int ResultCode, bool isServer)
+{
+	if (OnCompleted)
+	{
+		AsyncTask(ENamedThreads::GameThread, [this, ResultCode, Op]
+		{
+			OnCompleted(ResultCode, Op);
+		});
+	}
+}