+++ CHANGELOG.md
@@ -8,9 +8,18 @@
 ### Added
 
 - `UBeamCli` updated API bindings and performance improvements (reflected in Content Window responsiveness).
+- New and updated SDK install/update flow (Quick Start Docs updated to match this).
+  - Handles copying the plugin files as needed.
+  - Handles adding `Beam` utility code to your project's `Target.cs` file.
+  - Handles adding Beamable plugins to your `uproject` file.
+- `UBeamInventorySubsystem` now exposes `TryGetAllCurrenciesByGamerTag` and other similar functions to query the local state of inventories other than of one of the UserSlots.
+- `UBeamAnnouncementSubsystem` added to query and receive Beamable announcement notifications. 
+- Added `BEAMPROJ_LiveOpsDemo` to showcase usage of Beamable's LiveOps Platform (Stats, Inventory, Announcements).
+- Added `BEAMPROJ_DiscordDemo` to showcase usage of Beamable's Microservice to integrate with Discord for creating community management tools.
+- Added `BEAMPROJ_SteamDemo` to showcase usage of Beamable's Microservice and Identify federation to implement Steam Login.
 
 ### Changed
-
+- Updated Beamable CLI version to 2.1.0.
 - Visual feedback improvements to the HathoraDemo sample project:
   - Feedback for waiting for match start.
   - Feedback for victory/defeat screen.